fix(mcp-apps): address code review feedback
- Fix doc comment date to match PROTOCOL_VERSION constant (2025-03-26)
- Fix token masking to fully obscure short tokens (< 8 chars)
- Cast appIndex count to int for JS output safety
- Fix PHPCS alignment in REST controller

Agent-Logs-Url: https://github.com/nvdigitalsolutions/mcp-ai-wpoos/sessions/28dc5da2-3ee5-4910-97be-676d4069968f

+++ includes/assistants/metaboxes/class-wp-mcp-ai-metabox-mcp-apps.php
@@ -414,7 +414,7 @@ class WP_MCP_AI_Metabox_MCP_Apps extends WP_MCP_AI_Metabox_Base {
 		?>
 		<script type="text/javascript">
 		( function() {
-			var appIndex = <?php echo count( $this->get_current_apps_count() ); ?>;
+			var appIndex = <?php echo (int) count( $this->get_current_apps_count() ); ?>;
 			var maxApps = 10;
 
 			document.addEventListener( 'DOMContentLoaded', function() {
